Introduce worldUtil
